Revert "Commit TEST validé"
This reverts commit a5dcb4e1901177292f73bb75333098721aa3cf10.

+++ ProtocoleC/mathC.h
@@ -1,12 +1,10 @@
 #ifndef DEF_MATH
 #define DEF_MATH
 
-#include "mathC.h"
-
 typedef struct vector3 {
-	int x;
-	int y;
-	int z;
+	float x;
+	float y;
+	float z;
 } Vector3;
 
 int minValue(int a, int b);
